feat(resourcemanager): use a unified background metrics flusher
